+++ R/package.R
@@ -1,7 +1,3 @@
-#' multiverse.internals: Internal Infrastructure for R-multiverse.
-#' @description Internal Infrastructure for R-multiverse.
-#' @name multiverse.internals-package
-#' @family help
 #' @importFrom gh gh
 #' @importFrom jsonlite parse_json read_json stream_in write_json
 #' @importFrom nanonext ncurl parse_url status_code